update location changes

+++ map/views.py
@@ -1,7 +1,11 @@
 from django.shortcuts import render, redirect
 from .models import Map
+from cars.models import Cars
+from .forms import MapForm
+from cars.forms import CarForm
 from .forms import MapForm
 from django.contrib import messages
+from cars.views import list_cars
 
 from django.contrib.auth.decorators import login_required
 
@@ -30,12 +34,12 @@ def list_map(request):
 """Update Location"""
 
 
-def show_map(request, site_id):
+def update_map(request, site_id):
     try:
         site = Map.objects.get(pk=site_id)
     except Map.DoesNotExist:
         messages.error(request, 'There is no location with this id!')
-        return redirect('map')
+        return redirect(list_cars)
     map_form = MapForm(initial={
         'plz': site.plz,
         'location': site.location,
@@ -44,21 +48,24 @@ def show_map(request, site_id):
     })
     if request.method == 'POST':
         map_form = MapForm(request.POST, instance=site)
-    if map_form.is_valid():
-        site.plz = map_form.cleaned_data['plz']
-        site.location = map_form.cleaned_data['location']
-        site.address = map_form.cleaned_data['address']
-        site.country = map_form.cleaned_data['country']
-        site.save()
-    messages.success(request, 'Location successfully updated.')
-    return render('functions/map.html', {'map_form': map_form, 'site': site})
+        if map_form.is_valid():
+            site.plz = map_form.cleaned_data['plz']
+            site.location = map_form.cleaned_data['location']
+            site.address = map_form.cleaned_data['address']
+            site.country = map_form.cleaned_data['country']
+            site.save()
+        messages.success(request, 'Location successfully updated.')
+    return redirect('functions/list_cars', {'map_form': map_form})
 
 
 def delete_map(request, site_id):
     try:
         site = Map.objects.get(pk=site_id)
-        site.delete()
-        messages.success(request, 'Location successfully deleted.')
+        if Cars.objects.filter(location_id=site_id):
+            messages.error(request, 'This site is still linked with a car.')
+        else:
+            site.delete()
+            messages.success(request, 'Location successfully deleted.')
     except Map.DoesNotExist:
         messages.error(request, 'There is no location with this id.')
-    return redirect('map')
+    return redirect('list_cars')
